requirements_notifications.txt
# Push Notifications
firebase-admin==6.2.0
django-model-utils==4.3.1
